# Hoist object allocation before inner field initialization