@gbaraldi 2022-04-27
`abstractarrays` failure minized to 
```julia
function foo()
    for sz in ((5, 3), (7, 11))
        for idxs in ((1:sz[1], 1:sz[2]), (1:sz[1], 2:2:sz[2]),)
        end
    end
end
foo()
```